backend/requirements.txt
fastapi
uvicorn
pdfplumber
pytesseract
pillow
python-multipart
python-magic
spacy
nameparser
pandas
